LPS-146975 Moves code to the class where we are using it, since it is a component
